requirements.txt
python-dotenv
streamlit
streamlit-chat
google-generativeai
ipython